+++ 최종(웹)/web/myapp/templates/navbar.html
@@ -17,7 +17,7 @@
                     <ul role="navigation" lang="en" aria-label="상단 메뉴">
 
                        
-                        <li><a href="{% url 'upload' %}">검색</a></li>
+                        <li><a href="{% url 'service' %}">검색</a></li>
                         <li><a href="{% url 'contact' %}">문의하기</a></li>
 
                         <li class="nav-item droopdown"><a class="nav-link dropdown-toggle" id="navbardrop" data-toggle="dropdown">커뮤니티</a>
@@ -82,7 +82,7 @@
                 <nav class="gnb">
                     <ul role="navigation" lang="en" aria-label="상단 메뉴">
                        
-                        <li><a href="{% url 'upload' %}">검색</a></li>
+                        <li><a href="{% url 'service' %}">검색</a></li>
                         <li><a href="{% url 'contact' %}">문의하기</a></li>
                         <li class="nav-item droopdown"><a class="nav-link dropdown-toggle" id="navbardrop" data-toggle="dropdown">커뮤니티</a>
                             <div class="dropdown-menu">
